Installed redux dev tool Moved App.js to components folder Corrected path errors in index.js
