+++ package.json
@@ -12,6 +12,7 @@
     "react-redux": "^7.2.4",
     "react-scripts": "4.0.3",
     "redux": "^4.1.0",
+    "redux-devtools-extension": "^2.13.9",
     "web-vitals": "^1.1.1"
   },
   "scripts": {
